resolving breaking changes post code cleaning
