+++ components/Dashboard.js
@@ -10,13 +10,15 @@ import ConnectWalletCustom from "./ConnectWalletCustom";
 
 import Image from "next/image";
 
-import { Framework } from "@superfluid-finance/sdk-core";
+import { apolloClient } from "@/helpers/apollo";
+import { truncateAddress } from "@/helpers/formatHelper";
+import { gql } from "@apollo/client";
 import { ethers } from "ethers";
 import ChainSelect from "./ChainSelect";
-import { apolloClient } from "@/helpers/apollo";
-import { gql } from "@apollo/client";
-import { truncateAddress } from "@/helpers/formatHelper";
 import DashboardRow from "./DashboardRow";
+
+// custom hooks
+import useXStream from "@/hooks/xStream/useXStream";
 
 function Dashboard() {
   const { address, isConnected } = useAccount();
@@ -28,6 +30,9 @@ function Dashboard() {
   const [dropDownAll, setDropDownAll] = useState(false);
   const [dropDownIncoming, setDropDownIncoming] = useState(true);
   const [dropDownOutgoing, setDropDownOutgoing] = useState(true);
+
+  // custom hooks
+  const hookXStream = useXStream();
 
   //integration
   const [allData, setAllData] = useState([]);
@@ -176,39 +181,7 @@ function Dashboard() {
     setAllData(allData);
   };
 
-  const getBalance = async () => {
-    try {
-      const { ethereum } = window;
-      if (ethereum) {
-        const provider = new ethers.providers.Web3Provider(ethereum);
-        const signer = provider.getSigner();
-        const account = await signer.getAddress();
-
-        const sf = await Framework.create({
-          chainId: 5,
-          provider: provider,
-        });
-
-        const DAIxContract = await sf.loadSuperToken(
-          "0x3427910EBBdABAD8e02823DFe05D34a65564b1a0"
-        );
-        const DAIx = DAIxContract.address;
-
-        try {
-          const b = await DAIxContract.balanceOf({
-            account: account,
-            providerOrSigner: signer,
-          });
-          const bal = ethers.utils.formatEther(b);
-          setBalane(bal);
-        } catch (error) {
-          console.error(error);
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  
 
   useEffect(() => {
     setDropDownAll(true);
@@ -270,7 +243,7 @@ function Dashboard() {
   }
 
   useEffect(() => {
-    getBalance();
+    hookXStream.getBalance();
     loadData();
   }, [address, chain]);
